requirements.txt
# Core Framework
streamlit>=1.28.0
websockets>=11.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Visualization
plotly>=5.17.0

# Analytics & Statistics
scipy>=1.11.0
statsmodels>=0.14.0
scikit-learn>=1.3.0
pykalman>=0.9.5

# Utilities
python-dateutil>=2.8.0